+++ src/app/state/tracking_view_state.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <array>
+#include <cstdint>
 
 enum class AngularMomentumViewMode {
   FaceOn = 0,
@@ -8,14 +9,14 @@ enum class AngularMomentumViewMode {
 
 struct TrackingTargetState {
   bool followParticle = false;
-  int targetParticleID = -1;
+  int64_t targetParticleID = -1;
 
   bool followClump = false;
   int targetClumpID = -1;
 
   bool followSinkParticle = false;
   bool followSinkParticleMostMassive = true;
-  int targetSinkParticleID = -1;
+  int64_t targetSinkParticleID = -1;
 
   bool useMassCenter = false;
   float massCenterRadius = 0.0f;       // <= 0 means unlimited.
